pyproject.toml
[tool.poetry]
name = "promptmage"
version = "0.1.2"
description = "\"PromptMage\" is designed to offer an intuitive interface that simplifies the process of creating and managing LLM workflows as a self-hosted solution."
authors = ["Tobias Sterbak <github@tobiassterbak.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://www.promptmage.io"
repository = "https://github.comm/tsterbak/promptmage"
keywords = ["promptmage", "llm", "workflow", "management", "self-hosted", "solution", "ai", "nlp", "prompt"]
classifiers = [
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
include = ["promptmage/static"]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
click = "^8.1.7"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
setuptools = "^69.5.1"
nicegui = "^1.4.25"
fastapi = ">=0.109.1,<0.110.0"
sqlalchemy = "^2.0.31"
python-slugify = "^8.0.4"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.2"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.23"
material-plausible-plugin = "^0.2.0"
pytest-selenium = "^4.1.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
promptmage = "promptmage.cli:promptmage"
